feat: 파일안에 여러 컴포넌트 있는 것들 분리 작업
